requirements.txt
fastapi
pydantic
uvicorn
pandas
numpy
scikit-learn
joblib
nltk
openai
qdrant-client
python-dotenv
matplotlib
